# Fix BeepwearPRO 1.03 not recognizing this device as a working Notebook Adapter