--- a/led_blaster.h
+++ b/led_blaster.h
@@ -5,6 +5,7 @@
 
 namespace LedBlaster {
   void setup();
+  void enable_fast_mode(bool fast_mode_enabled);
   void emit_0();
   void emit_1();
   void emit_byte(uint8_t serial_byte);