--- a/notebook_adapter.h
+++ b/notebook_adapter.h
@@ -4,8 +4,6 @@
 #include <Arduino.h>
 
 namespace NotebookAdapter {
-  void emulate_command_mode(uint8_t serial_byte);
-  void emulate_data_mode(uint8_t serial_byte);
   void emulate(uint8_t serial_byte);
 }