Fixes https://github.com/synthead/timex-datalink-arduino/issues/9! :sparkles: 

This PR fixes an issue where the BeepwearPRO 1.03 software doesn't recognize this emulator as a real Notebook Adapter.  Changes include:

- Recognizing the "V" command to switch to a faster data mode
- Echoing all received packets back to the serial host, namely for the "V" command
- Using "107\r M764 rev 764100" as the query payload

When the adapter sees the "V" command, it switches the adapter to a faster data mode of 6 ms illuminated, 408 ms pause that the Motorola Beepwear Pro must use for it to work.  These timings were measured using a logic analyzer with the original Timex Notebook Adapter's LED while performing a transfer with the BeepwearPRO 1.03 software.

With the current 15 ms illuminated, 450 ms pause timing, the Beepwear Pro will sync, but when it starts to parse data packets, it will display "COMM FAILED" and emits an audible siren effect.  The new "fast mode" swaps the adapter to use timings that are appropriate for the Beepwear Pro, and mitigates this issue completely while maintaining backward-compatibility for all other Timex Datalink devices.

The faster data mode is only enabled while the transmission occurs.   After the 1 second timeout, the data mode is set back to normal speed, and command mode is entered.  This works well because the "V" command (0x56) occurs after the initial handshakes to check if the serial device is the Notebook Adapter, and before the data mode (triggered by "U", or 0x55).

Here is an example of truncated bytes sent to the Notebook Adapter from the BeepwearPRO software:

```
78 3F 78 56 55 55 55 55 55 55 55 55 55 55 55 55 ...
```

In addition, the query command was changed from "?126\r M764 rev 764002" to "?107\r M764 rev 764100".  This response was sniffed from this setup:

- A logic analyzer with a MAX232N sniffing an original Timex Notebook Adapter
- A physical Windows 98 SE machine with a physical RS232C UART serial port
- Using the original BeepwearPRO 1.03 software

These changes were tested working with the following software:

- Timex BeepwearPRO 1.03
- Timex Data Link® Software 2.1d
- Timex Ironman Triathlon® Data Link® 2.01
- DSI Electronics e-BRAIN v1.1.6
- timex_datalink_client Ruby library

These devices were tested working with these changes:

- Motorola Beepwear Pro
- Timex Datalink 70
- Timex Ironman Triathlon

Before this PR:

![image](https://user-images.githubusercontent.com/820984/220309326-943f1659-5d90-470d-85e0-f9e072fad857.png)

With the changes in this PR:

![image](https://user-images.githubusercontent.com/820984/220309274-b955a833-d375-4abd-b11f-8cf1563845e6.png)